+++ corehq/blobs/tests/test_migrate.py
@@ -8,7 +8,7 @@ from corehq.blobs.mixin import BlobMixin
 from corehq.blobs.s3db import maybe_not_found
 from corehq.blobs.tests.util import (TemporaryFilesystemBlobDB,
     TemporaryMigratingBlobDB, TemporaryS3BlobDB)
-from corehq.util.couch_doc_processor import ResumableDocsByTypeIterator
+from corehq.util.doc_processor import ResumableDocsByTypeIterator
 from corehq.util.test_utils import trap_extra_setup
 
 from django.conf import settings
